moved collaboration out of planning, cleanup
