+++ src/planscape/planning/models.py
@@ -5,9 +5,7 @@ from django.conf import settings
 from utils.uuid_utils import generate_short_uuid
 from django.core.serializers.json import DjangoJSONEncoder
 from core.models import CreatedAtMixin, UpdatedAtMixin
-from django.contrib.contenttypes.models import ContentType
-from django.utils.translation import gettext_lazy as _
-import uuid
+import uuid, shortuuid
 
 User = get_user_model()
 
@@ -138,71 +136,3 @@ class SharedLink(CreatedAtMixin, UpdatedAtMixin, models.Model):
 
     class Meta:
         ordering = ["-created_at", "user"]
-
-class Role(models.TextChoices):
-    CREATOR = "Creator"
-    OWNER = 'Owner'
-    COLLABORATOR = 'Collaborator'
-    VIEWER = 'Viewer'
-    
-class Permissions(models.Model):
-    role = models.CharField(
-        choices=Role.choices,
-        max_length=16,
-        default=Role.VIEWER,
-    )
-    permission = models.CharField(max_length=60)
-    class Meta:
-        constraints = [
-            # a person can only be invited once to a specific planning area
-            models.UniqueConstraint(
-                fields=[
-                    "role",
-                    "permission"
-                ],
-                name="unique_permission",
-            )
-        ]
-
-class Collaborator(models.Model):
-    uuid = models.UUIDField(default=uuid.uuid4, unique=True)
-    # the email address invited to collaborate
-    email = models.CharField(max_length=120)
-    # the user that is being added as collaborator
-    # might be empty if no user is found with the email
-    collaborator = models.ForeignKey(
-        User,
-        related_name="collaborator",
-        on_delete=models.CASCADE,
-        null=True,
-    )
-    # the role assigned 
-    role = models.CharField(
-        choices=Role.choices,
-        max_length=16,
-        default=Role.VIEWER,
-    )
-    # the user that invited the collaborator
-    inviter = models.ForeignKey(
-        User,
-        related_name="inviter",
-        on_delete=models.CASCADE,
-        null=False,
-    )
-    # use content types to potentially save other things that are not only planning_areas
-    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_pk = models.CharField(_('object ID'), max_length=255)
-
-    class Meta:
-        constraints = [
-            # a person can only be invited once to a specific planning area
-            models.UniqueConstraint(
-                fields=[
-                    "email",
-                    "content_type",
-                    "object_pk",
-                ],
-                name="unique_collaborator",
-            )
-        ]
-
